fix(cardeditor): fixed Open JSON editor button in config pane
