+++ packages/react/src/components/CardEditor/CardEditForm/CardEditForm.jsx
@@ -1,13 +1,9 @@
-import React, { useState, useMemo } from 'react';
+import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
-import { Code16 } from '@carbon/icons-react';
-import { isEmpty, omit, pick } from 'lodash-es';
 
 import { CARD_DIMENSIONS, CARD_TYPES } from '../../../constants/LayoutConstants';
 import { settings } from '../../../constants/Settings';
-import Button from '../../Button';
 import { Tabs, Tab } from '../../Tabs';
-import CardCodeEditor from '../../CardCodeEditor/CardCodeEditor';
 import { DataItemsPropTypes } from '../../DashboardEditor/editorUtils';
 
 import CardEditFormContent from './CardEditFormContent';
@@ -61,22 +57,12 @@ const propTypes = {
    * this prop will be ignored if getValidDataItems is defined
    */
   dataItems: DataItemsPropTypes,
-  /** if provided, allows the consumer to make changes to the cardConfig for preview in the JSON editor modal.
-   * onCardJsonPreview(card)
-   */
-  onCardJsonPreview: PropTypes.func,
   /** an object where the keys are available dimensions and the values are the values available for those dimensions
    *  ex: { manufacturer: ['Rentech', 'GHI Industries'], deviceid: ['73000', '73001', '73002'] }
    */
   availableDimensions: PropTypes.shape({}),
-  /** If provided, runs the function when the user clicks submit in the Card code JSON editor
-   * onValidateCardJson(cardConfig)
-   * @returns Array<string> error strings. return empty array if there is no errors
-   */
-  onValidateCardJson: PropTypes.func,
   currentBreakpoint: PropTypes.string,
   isSummaryDashboard: PropTypes.bool,
-  testID: PropTypes.string,
   /** optional link href's for each card type that will appear in a tooltip */
   dataSeriesItemLinks: PropTypes.shape({
     simpleBar: PropTypes.string,
@@ -121,13 +107,10 @@ const defaultProps = {
   },
   getValidDataItems: null,
   getValidTimeRanges: null,
-  onCardJsonPreview: null,
   dataItems: [],
   availableDimensions: {},
-  onValidateCardJson: null,
   currentBreakpoint: 'xl',
   isSummaryDashboard: false,
-  testID: 'card-edit-form',
   dataSeriesItemLinks: null,
 };
 
@@ -143,131 +126,22 @@ export const getCardSizeText = (size, i18n) => {
   return `${sizeName} ${sizeDimensions}`;
 };
 
-/**
- * Returns errors of basic JSON syntax
- * Must not be empty string, Must be valid JSON
- * @param {Object} card JSON currently being edited
- * @returns {Array<string>} error strings
- */
-export const basicCardValidation = (card) => {
-  const errors = [];
-  try {
-    const json = JSON.parse(card);
-    if (!json || typeof json !== 'object') {
-      errors.push(`${card.substring(0, 8)} is not valid JSON`);
-    }
-  } catch (e) {
-    errors.push(e.message);
-  }
-  return errors;
-};
-
-/**
- * removes properties needed for the editor that we don't want the user to be able to modify
- * @param {Object} card JSON currently being edited
- * @returns {Object} card with omitted attributes
- */
-export const hideCardPropertiesForEditor = (card) => {
-  let attributes;
-  let series;
-  let columns;
-  if (card.content?.attributes) {
-    attributes = card.content.attributes.map((attribute) =>
-      omit(attribute, ['aggregationMethods', 'grain', 'downSampleMethods'])
-    );
-  }
-  if (card.content?.series) {
-    series = card.content.series.map((attribute) =>
-      omit(attribute, ['aggregationMethods', 'grain', 'downSampleMethods'])
-    );
-  }
-  if (card.content?.columns) {
-    columns = card.content.columns.map((column) =>
-      omit(column, ['aggregationMethods', 'grain', 'downSampleMethods'])
-    );
-  }
-  // Need to exclued content for custom cards because the card's JSX element lives on it in this case
-  if (!CARD_TYPES.hasOwnProperty(card.type) || card.type === CARD_TYPES.CUSTOM) {
-    return omit(card, 'content');
-  }
-  return omit(
-    attributes // VALUE CARD
-      ? { ...card, content: { ...card.content, attributes } }
-      : series // TIMESERIES AND BAR CHART CARDS
-      ? { ...card, content: { ...card.content, series } }
-      : card.values?.hotspots // IMAGE CARD
-      ? {
-          ...card,
-          values: {
-            ...card.values,
-            hotspots: card.values?.hotspots?.map((hotspot) => ({
-              ...hotspot,
-              content: {
-                ...hotspot.content,
-                attributes: hotspot.content?.attributes?.map((attribute) =>
-                  omit(attribute, ['aggregationMethods', 'grain'])
-                ),
-              },
-            })),
-          },
-        }
-      : columns // TABLE CARD
-      ? { ...card, content: { ...card.content, columns } }
-      : card,
-    ['content.src', 'content.imgState', 'i18n', 'validateUploadedImage']
-  );
-};
-
-/**
- * Checks for JSON form errors
- * @param {Object} card JSON text input
- * @param {Function} setError
- * @param {Function} onValidateCardJson
- * @param {Function} onChange
- * @param {Function} setShowEditor
- */
-export const handleSubmit = (card, id, setError, onValidateCardJson, onChange, setShowEditor) => {
-  // first validate basic JSON syntax
-  const basicErrors = basicCardValidation(card);
-  // second validate the consumer's custom function if provided
-  let customValidationErrors = [];
-  if (onValidateCardJson) {
-    customValidationErrors = onValidateCardJson(card);
-  }
-  const allErrors = basicErrors.concat(customValidationErrors);
-  // then submit
-  if (isEmpty(allErrors)) {
-    onChange({ ...JSON.parse(card), id });
-    setShowEditor(false);
-    return true;
-  }
-
-  setError(allErrors.join('. '));
-  return false;
-};
-
 const CardEditForm = ({
   cardConfig,
   isSummaryDashboard,
   onChange,
   i18n,
   dataItems,
-  onValidateCardJson,
-  onCardJsonPreview,
   getValidDataItems,
   getValidTimeRanges,
   currentBreakpoint,
   availableDimensions,
-  testID,
   dataSeriesItemLinks,
   // eslint-disable-next-line react/prop-types
   onFetchDynamicDemoHotspots,
 }) => {
   const mergedI18n = useMemo(() => ({ ...defaultProps.i18n, ...i18n }), [i18n]);
-  const [showEditor, setShowEditor] = useState(false);
-  const [modalData, setModalData] = useState();
 
-  const { id } = cardConfig;
   const baseClassName = `${iotPrefix}--card-edit-form`;
 
   const isCustomCardWithNoSettings =
@@ -275,77 +149,36 @@ const CardEditForm = ({
     !cardConfig.renderEditSettings;
 
   return (
-    <>
-      {showEditor ? (
-        <CardCodeEditor
-          onSubmit={(card, setError) =>
-            handleSubmit(card, id, setError, onValidateCardJson, onChange, setShowEditor)
-          }
-          onClose={() => setShowEditor(false)}
-          initialValue={modalData}
-          i18n={pick(
-            mergedI18n,
-            'errorTitle',
-            'modalTitle',
-            'modalLabel',
-            'modalHelpText',
-            'modalIconDescription',
-            'copyBtnDescription',
-            'copyBtnFeedBack',
-            'expandBtnLabel',
-            'modalPrimaryButtonLabel',
-            'modalSecondaryButtonLabel'
-          )}
-        />
-      ) : null}
-      <div className={baseClassName}>
-        <Tabs scrollIntoView={false}>
-          <Tab label={mergedI18n.contentTabLabel}>
-            <CardEditFormContent
+    <div className={baseClassName}>
+      <Tabs scrollIntoView={false}>
+        <Tab label={mergedI18n.contentTabLabel}>
+          <CardEditFormContent
+            cardConfig={cardConfig}
+            onChange={onChange}
+            isSummaryDashboard={isSummaryDashboard}
+            i18n={mergedI18n}
+            dataItems={dataItems}
+            availableDimensions={availableDimensions}
+            getValidDataItems={getValidDataItems}
+            getValidTimeRanges={getValidTimeRanges}
+            currentBreakpoint={currentBreakpoint}
+            dataSeriesItemLinks={dataSeriesItemLinks}
+            onFetchDynamicDemoHotspots={onFetchDynamicDemoHotspots}
+          />
+        </Tab>
+        {!isCustomCardWithNoSettings ? (
+          <Tab label={mergedI18n.settingsTabLabel}>
+            <CardEditFormSettings
+              availableDimensions={availableDimensions}
               cardConfig={cardConfig}
               onChange={onChange}
-              isSummaryDashboard={isSummaryDashboard}
               i18n={mergedI18n}
-              dataItems={dataItems}
-              availableDimensions={availableDimensions}
               getValidDataItems={getValidDataItems}
-              getValidTimeRanges={getValidTimeRanges}
-              currentBreakpoint={currentBreakpoint}
-              dataSeriesItemLinks={dataSeriesItemLinks}
-              onFetchDynamicDemoHotspots={onFetchDynamicDemoHotspots}
             />
           </Tab>
-          {!isCustomCardWithNoSettings ? (
-            <Tab label={mergedI18n.settingsTabLabel}>
-              <CardEditFormSettings
-                availableDimensions={availableDimensions}
-                cardConfig={cardConfig}
-                onChange={onChange}
-                i18n={mergedI18n}
-                getValidDataItems={getValidDataItems}
-              />
-            </Tab>
-          ) : null}
-        </Tabs>
-        <div className={`${baseClassName}--footer`}>
-          <Button
-            testId={`${testID}-open-editor-button`}
-            kind="ghost"
-            size="small"
-            renderIcon={Code16}
-            onClick={() => {
-              const cardConfigForModal = onCardJsonPreview
-                ? onCardJsonPreview(hideCardPropertiesForEditor(cardConfig))
-                : hideCardPropertiesForEditor(cardConfig);
-              setModalData(JSON.stringify(cardConfigForModal, null, 4));
-              setShowEditor(true);
-            }}
-          >
-            {mergedI18n.openEditorButton}
-          </Button>
-        </div>
-      </div>
-    </>
+        ) : null}
+      </Tabs>
+    </div>
   );
 };
